requirements.txt
Pillow~=9.1.0
PyPDF2~=1.27.9